update bank and wallet

+++ app/src/main/java/ww/smartexpress/app/ui/payout/PayoutViewModel.java
@@ -18,7 +18,7 @@ import ww.smartexpress.app.ui.base.activity.BaseViewModel;
 public class PayoutViewModel extends BaseViewModel {
 
     public ObservableField<String> money = new ObservableField<>();
-    public ObservableField<Integer> balance = new ObservableField<>(0);
+    public ObservableField<Double> balance = new ObservableField<>(0.0);
     public ObservableField<UserEntity> user = new ObservableField<>();
     public ObservableField<BankCard> bankCard = new ObservableField<>();
     public PayoutViewModel(Repository repository, MVVMApplication application) {
@@ -37,7 +37,7 @@ public class PayoutViewModel extends BaseViewModel {
             showErrorMessage("Số tiền rút tối thiểu là 50.000đ");
             return;
         }
-        if(money.get() != null && Integer.valueOf(money.get())> balance.get()){
+        if(money.get() != null && Double.valueOf(money.get()) > balance.get()){
             showErrorMessage("Số tiền rút vượt quá số dư trong ví");
             return;
         }
